Move /scheduler/ into /cmd/scheduler/
This is to conform more closely to Golang project structure and to make
the code easier to read and extend.
