+++ cmd/scheduler/Dockerfile
@@ -1,7 +1,7 @@
 FROM golang@sha256:3c4de86eec9cbc619cdd72424abd88326ffcf5d813a8338a7743c55e5898734f as build
 WORKDIR /src
 COPY . ./
-RUN go build -o scheduler main.go
+RUN go build -o scheduler ./cmd/scheduler/main.go
 
 
 FROM gcr.io/distroless/base:nonroot@sha256:bc84925113289d139a9ef2f309f0dd7ac46ea7b786f172ba9084ffdb4cbd9490
