+++ cmd/scheduler/main.go
@@ -13,7 +13,7 @@ import (
 	_ "gocloud.dev/pubsub/gcppubsub"
 	_ "gocloud.dev/pubsub/kafkapubsub"
 
-	"github.com/ossf/package-analysis/scheduler/proxy"
+	"github.com/ossf/package-analysis/cmd/scheduler/proxy"
 	"github.com/ossf/package-feeds/pkg/feeds"
 )
